[stack-manager] Add ImportStackProjectAction
